♻️ refactor [#11.4.4]: 3차 개선 - 벤치마크 엔진 모듈화 및 에러 가시성 확보
- **[tests/performance]**:
  - [__init__.py](tests/performance/__init__.py:0:0-0:0) 추가를 통한 패키지화.
  - [collector.py](tests/performance/collector.py:0:0-0:0) 신설: 지표 측정 엔진(measure_stream_performance) 이동 및 고도화.
  - `sys.path.insert` 해킹 제거 및 절대 패키지 임포트 적용.
- **[benchmark_rag.py]**:
  - 부하 테스트 중 발생한 에러 요약(Exception 개수 및 샘플) 로깅 추가.
  - 논리적 실패(Logical failures) 집계 기능 강화.
- **[collector.py]**:
  - JSON 파싱 오류 시 건너뛰기 로직(Robustness) 및 0으로 나누기 방어 적용.

🔗 Related:
- Issue [#614]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/708#pullrequestreview-3944779569)
