+++ tests/performance/__init__.py
@@ -1,0 +1,1 @@
+# tests/performance/__init__.py
